+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
    `SplittingStrategy` in component fusion packs.
 - added equation generators to `FusionStrategy` for the case where one or both
    sides are positive
+- added a `SubclassVerification` strategy
 
 ### Changed
 - insertion packs now use the `one_cell_only` option, and no longer use